requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
arpeggio==1.10.2
    # via parver
attrs==21.2.0
    # via parver
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.6
    # via requests
dill==0.3.4
    # via pulumi
grpcio==1.41.0
    # via pulumi
idna==3.2
    # via requests
parver==0.3.1
    # via
    #   pulumi-kubernetes
    #   pulumi-minio
protobuf==3.18.0
    # via pulumi
pulumi==3.13.2
    # via
    #   -r requirements.in
    #   pulumi-kubernetes
    #   pulumi-minio
pulumi-kubernetes==3.7.3
    # via -r requirements.in
pulumi-minio==0.1.0
    # via -r requirements.in
python-libtorrent-bin==2.0.2
    # via -r requirements.in
pyyaml==5.4.1
    # via
    #   pulumi
    #   pulumi-kubernetes
requests==2.26.0
    # via
    #   pulumi-kubernetes
    #   transmission-rpc
semver==2.13.0
    # via
    #   pulumi
    #   pulumi-kubernetes
    #   pulumi-minio
six==1.16.0
    # via
    #   grpcio
    #   parver
    #   pulumi
transmission-rpc==3.3.0
    # via -r requirements.in
typing-extensions==3.10.0.2
    # via transmission-rpc
urllib3==1.26.7
    # via requests
